Namespace adjustment
