+++ TBS.Data/Models/Post/Shipper/ShipperPost.cs
@@ -1,31 +1,29 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
+using TBS.Data.Models.General;
 using TBS.Data.Models.Vehicle;
 
-namespace TBS.Data.Models.Post
+namespace TBS.Data.Models.Post.Shipper
 {
-    // Transporter post
-    public class CarrierPost
+    // Dealer post
+    public class ShipperPost
     {
         public int Id { get; set; }
 
-        [Required(ErrorMessage = "Pickup city required")]
-        public string PickupCity { get; set; }
+        [Required(ErrorMessage = "Vehicle required")]
+        public PostedVehicle Vehicle { get; set; }
 
-        [Required(ErrorMessage ="Pickup date required")]
+        [Required(ErrorMessage = "Pickup location required")]
+        public Address PickupLocation { get; set; }
+
+        [Required(ErrorMessage = "Pickup date required")]
         public DateTime PickupDate { get; set; }
 
-        [Required(ErrorMessage = "Dropoff city required")]
-        public string DropoffCity { get; set; }
+        [Required(ErrorMessage = "Dropoff location required")]
+        public Address DropoffLocation { get; set; }
 
         [Required(ErrorMessage = "Dropoff date required")]
         public DateTime DropoffDate { get; set; }
-
-        [Required(ErrorMessage = "Trailer type required")]
-        public CarrierTrailerType TrailerType { get; set; }
-
-        [Required(ErrorMessage = "Spaces available required")]
-        public int SpacesAvailable { get; set; }
 
         [Required(ErrorMessage = "Cost required")]
         [Range(0, double.MaxValue, ErrorMessage = "Must be a valid cost")]
