+++ TBS.Data/Models/Vehicle/Carrier/CarrierTrailerType.cs
@@ -1,4 +1,4 @@
-﻿namespace TBS.Data.Models.Vehicle
+﻿namespace TBS.Data.Models.Vehicle.Carrier
 {
     public enum CarrierTrailerType
     {
